requirements.txt
fastapi
uvicorn
supabase
python-multipart
pydantic-settings
python-dotenv
email-validator
pytest
pytest-asyncio
httpx
